+++ domain/usecase/src/main/java/com/jusicool/usecase/market/GetTotalRecommendMarketListUseCase.kt
@@ -1,6 +1,8 @@
 package com.jusicool.usecase.market
 
 import com.jusicool.entity.market.RecommendMarketWithPrice
+import com.jusicool.usecase.market.util.SortOption
+import com.jusicool.usecase.market.util.sortedByOption
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
 import javax.inject.Inject
@@ -19,18 +21,4 @@ class GetTotalRecommendMarketListUseCase @Inject constructor(
             (stocks + cryptos).sortedByOption(sortBy)
         }
     }
-}
-
-enum class SortOption {
-    BY_PROFIT_DESC,
-    BY_PRICE_DESC,
-    BY_NAME_ASC,
-}
-
-private fun List<RecommendMarketWithPrice>.sortedByOption(
-    option: SortOption
-): List<RecommendMarketWithPrice> = when (option) {
-    SortOption.BY_PROFIT_DESC -> sortedByDescending { it.profitRate }
-    SortOption.BY_PRICE_DESC -> sortedByDescending { it.currentPrice }
-    SortOption.BY_NAME_ASC -> sortedBy { it.koreanName }
 }
